resplit execute tx relative function into writer executor
